Switch config to be indexed based on a task stream name
Using HITLayoutId as the way to look up HIT config is problematic
because it changes with each minor change to the template. So when
it changes, clients would need to change a) the S3 folder name, b)
the 'layouts' map key in config, c) the job that uploads CSVs to
S3. As such, it seemed better to use a user-defined stream name, for
which HITLayoutId was simply one of the config parameters. To
facilitate this, we've add a CloudFormation template to set up a
new HIT stream. The nice side-effect of this is that we can then
set up a SQS queue for the HIT output, instead of expecting the
user to do this manually.